backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.1
Pillow==8.3.1
faker==8.13.2
twilio==7.8.0
python-dotenv==0.19.0
geopy==2.2.0
Werkzeug==2.0.1
passlib==1.7.4
email-validator==1.1.3
supabase==1.0.3
python-jose==3.3.0
requests==2.28.1